# Speed up Cube.subset/Coord.intersect